requirements.txt
instaloader==4.9.6
requests==2.31.0
google-api-python-client==2.100.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
schedule==1.2.0
